refactor(*): update module/tests to point to new locations
