+++ src/ramstk/modules/function/Model.py
@@ -7,12 +7,10 @@
 """Function Package Data Model."""
 
 # Third Party Imports
-# Import third party packages.
 from pubsub import pub
 
 # RAMSTK Package Imports
-from ramstk.dao.programdb import RAMSTKFunction
-# Import other RAMSTK modules.
+from ramstk.models.programdb import RAMSTKFunction
 from ramstk.modules import RAMSTKDataModel
